+++ EOLib/Domain/Party/PartyActions.cs
@@ -14,20 +14,20 @@ namespace EOLib.Domain.Party
             _packetSendService = packetSendService;
         }
 
-        public void RequestParty(PartyRequestType type, short targetCharacterId)
+        public void RequestParty(PartyRequestType type, int targetCharacterId)
         {
             var packet = new PacketBuilder(PacketFamily.Party, PacketAction.Request)
-                .AddChar((byte)type)
+                .AddChar((int)type)
                 .AddShort(targetCharacterId)
                 .Build();
 
             _packetSendService.SendPacket(packet);
         }
 
-        public void AcceptParty(PartyRequestType type, short targetCharacterId)
+        public void AcceptParty(PartyRequestType type, int targetCharacterId)
         {
             var packet = new PacketBuilder(PacketFamily.Party, PacketAction.Accept)
-                .AddChar((byte)type)
+                .AddChar((int)type)
                 .AddShort(targetCharacterId)
                 .Build();
 
@@ -42,7 +42,7 @@ namespace EOLib.Domain.Party
             _packetSendService.SendPacket(packet);
         }
 
-        public void RemovePartyMember(short targetCharacterId)
+        public void RemovePartyMember(int targetCharacterId)
         {
             var packet = new PacketBuilder(PacketFamily.Party, PacketAction.Remove)
                 .AddShort(targetCharacterId)
@@ -54,12 +54,12 @@ namespace EOLib.Domain.Party
 
     public interface IPartyActions
     {
-        void RequestParty(PartyRequestType type, short targetCharacterId);
+        void RequestParty(PartyRequestType type, int targetCharacterId);
 
-        void AcceptParty(PartyRequestType type, short targetCharacterId);
+        void AcceptParty(PartyRequestType type, int targetCharacterId);
 
         void ListParty();
 
-        void RemovePartyMember(short targetCharacterId);
+        void RemovePartyMember(int targetCharacterId);
     }
 }
